fix: join channel flow in thread page
